refactor: move enum classes into enums module
