--- backend/application/enums/access_control_role.py
@@ -1,7 +1,0 @@
-import enum
-
-
-class AccessControlRole(enum.Enum):
-    USER = "USER"
-    ADMIN = "ADMIN"
-    MANAGER = "MANAGER"
