--- backend/application/enums/application_status.py
@@ -1,7 +1,0 @@
-import enum
-
-
-class ApplicationStatus(enum.Enum):
-    ACCEPTED = "ACCEPTED"
-    REJECTED = "REJECTED"
-    PENDING = "PENDING"
