--- backend/application/enums/role_status.py
@@ -1,7 +1,0 @@
-import enum
-
-
-class RoleStatus(enum.Enum):
-    OPEN = "OPEN"
-    CLOSED = "CLOSED"
-    PENDING = "PENDING"
