+++ backend/tests/conftest.py
@@ -1,18 +1,10 @@
 from flask.testing import FlaskClient
 import pytest
-from datetime import datetime
 
 from application.config import TestConfig
 from application import init_app
 from application.extensions import db
 from application import register_blueprints, register_error_handlers
-from application.models.skill import Skill
-from application.models.access_control import AccessControl
-from application.models.role import Role
-from application.models.staff import Staff
-from application.enums.access_control_role import AccessControlRole
-from application.models.role_listing import RoleListing
-from application.models.role_listing import time_format
 from sqlalchemy.orm import close_all_sessions
 from application import init_db
